10.16_리뷰답글 수정해야함

+++ src/main/java/com/example/ApiRound/crm/hyeonah/review/UserReviewDto.java
@@ -13,7 +13,7 @@ public class UserReviewDto {
     private String userName; // 사용자 이름
     private Long itemId;
     private String itemName; // 아이템 이름
-    private Integer bookingId;
+    private Long bookingId;  // reservations.id (Long)
     private Byte rating;
     private String title;
     private String content;
@@ -23,7 +23,7 @@ public class UserReviewDto {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     
-    // 답글 목록
+    // 답글 정보
     private List<UserReviewReplyDto> replies;
     
     // 통계
