+++ src/main/java/com/example/ApiRound/crm/hyeonah/review/UserReviewRepository.java
@@ -16,8 +16,8 @@ public interface UserReviewRepository extends JpaRepository<UserReview, Integer>
     // 특정 사용자의 리뷰 목록 조회
     List<UserReview> findByUserIdOrderByCreatedAtDesc(Integer userId);
     
-    // 특정 예약의 리뷰 조회
-    UserReview findByBookingId(Integer bookingId);
+    // booking_id로 리뷰 조회
+    UserReview findByBookingId(Long bookingId);
     
     // 특정 아이템의 평균 평점 조회
     @Query("SELECT AVG(r.rating) FROM UserReview r WHERE r.itemId = :itemId AND r.isPublic = true")
@@ -30,7 +30,7 @@ public interface UserReviewRepository extends JpaRepository<UserReview, Integer>
     @Query("SELECT COUNT(r) FROM UserReview r WHERE r.itemId = :itemId AND r.rating = :rating AND r.isPublic = true")
     Long countByItemIdAndRating(@Param("itemId") Long itemId, @Param("rating") Byte rating);
     
-    // 예약 ID로 리뷰 존재 여부 확인
-    boolean existsByBookingId(Integer bookingId);
+    // booking_id로 리뷰 존재 여부 확인
+    boolean existsByBookingId(Long bookingId);
 }
